Draws lines along pixel centres instead of edges for consistency
Rectangles and lines are aligned based on the centre of pixels (x + 0.5, y +
0.5). Drawing lines along the exact border of two pixels means that they will
round up or down in different ways on different machines. This results in the
horizontal lines being shifted a pixel down on my NVIDIA/Arch machine and
produces small gaps at he bottom of the fill. Drawing lines through the creter
of pixels removes the potential for rounding and prevents this bug.